swc-windows-installer.py: Set default logging level to INFO
And add the other levels as --verbose choices, even though we
currently only use LOG.info and LOG.debug.  Many *nix commands are
silent by default and have options for increasing the verbosity, but
Ethan points out that we should probably default to the setting that
makes the most sense for new users.