swc-windows-installer.py: Fetch SQLite over HTTP (not HTTPS)
This avoids issues with the user's certificate store [1] by skipping
TLS.  That means that middlemen will know you're asking for the shell
zip, but they'd probably be able to guess something similar from the
size of the response payload anyway.  All of our other fetches are
already over HTTP, and we don't have to worry about verifying the
server because we have cryptographically-strong hashes for verifying
the file they sent us.

[1]: https://github.com/swcarpentry/windows-installer/issues/41#issuecomment-215565310